Change file name to PascalCase format
