+++ src/webportal/src/app/user/fabric/components/MessageBox.jsx
